Fix a crash when setter is nullpta
In the current implementation, if we want to make a readonly property,
we should pass empty setter that is actual function but has no body
as follows:
  void HelloSetter(const Npai::CallbackInfo& info,
                   const Npai::Value& value) {
    // This is an empty setter to make a readonly property.
    // Do nothing
  }

  InstanceAccessor("hello_property", &HelloGetter, &HelloSetter);

If we allows taking `nullptr` as the argument, we can simplify user's
code more as follows:
  InstanceAccessor("hello_property", &HelloGetter, nullptr);

Currently, above code makes a crash. So, this patch makes
InstanceAccessor() take `nullptr` as the argument and then add some
tests for them.

PR-URL: https://github.com/nodejs/node-addon-api/pull/159
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>
Reviewed-By: Jason Ginchereau <jasongin@microsoft.com>